+++ packages/apps/plugins/plugin-chess/src/components/index.ts
@@ -6,3 +6,4 @@ import React from 'react';
 
 // Lazily load components for content surfaces.
 export const ChessMain = React.lazy(() => import('./ChessMain'));
+export const ChessArticle = React.lazy(() => import('./ChessArticle'));
